refactor: Create useLocation hook
